Handle unique partial indices properly
Do not treat columns with unique partial indexes as unique.  A unique
partial index enforces uniqueness across subset of column values, not
across the entire column.

This fixes a code generation error where, in the case where the
constrained column is a foreign key, the referenced table has
a one-to-one relationship to the indexed table and not one-to-many as
expected.